feat: membuat fitur send qr ke partisipan

+++ internal/domain/repository/organizer_repository.go
@@ -9,5 +9,5 @@ import (
 type OrganizerRepository interface {
 	Create(ctx context.Context, organizer *domain.Organizer) error
 	GetByEmail(ctx context.Context, email string) (*domain.Organizer, error)
-	GetByID(ctx context.Context, id int) (*domain.Organizer, error)
+	GetByID(ctx context.Context, id int64) (*domain.Organizer, error)
 }
